+++ crates/icydb-core/src/db/types/subaccount.rs
@@ -5,6 +5,10 @@ use crate::{
     types::Subaccount,
 };
 
+///
+/// Subaccount
+///
+
 impl Filterable for Subaccount {
     type Filter = NoFilterKind;
     type ListFilter = NoFilterKind;