+++ crates/icydb-core/src/db/types/timestamp.rs
@@ -8,6 +8,10 @@ use crate::{
     types::Timestamp,
 };
 
+///
+/// Timestamp
+///
+
 impl Filterable for Timestamp {
     type Filter = Nat64RangeFilterKind;
     type ListFilter = Nat64ListFilterKind;